Split core match semantics tests
